--- main/client/src/components/editEnvelope.js
@@ -1,13 +1,0 @@
-import React from "react";
-import EnvelopeForm from "./envelopeForm";
- 
-export default function EditEnvelope() {
- 
-  // This following section will display the form that takes input from the user to update the data.
-  return (
-    <div>
-      <h3>Update Envelope</h3>
-      <EnvelopeForm />
-    </div>
-  );
-}
